+++ src/renderer/src/App.tsx
@@ -22,6 +22,7 @@ import { Timeline } from './pages/Timeline'
 import { ExamPrep } from './pages/ExamPrep'
 import { MapView } from './pages/MapView'
 import { Intel } from './pages/Intel'
+import { InfraIntel } from './pages/InfraIntel'
 import { Vpn } from './pages/Vpn'
 import { WhatsNew } from './pages/WhatsNew'
 import { Settings } from './pages/Settings'
@@ -58,6 +59,7 @@ export default function App(): JSX.Element {
               <Route path="/exam-prep" element={<ExamPrep />} />
               <Route path="/map" element={<MapView />} />
               <Route path="/intel" element={<Intel />} />
+              <Route path="/infra" element={<InfraIntel />} />
               <Route path="/graph" element={<Graph />} />
               <Route path="/vpn" element={<Vpn />} />
               <Route path="/whats-new" element={<WhatsNew />} />